Slight change to handling temporary sync workers
Changes based on: https://github.com/rwmt/Multiplayer-Compatibility/pull/275#issuecomment-1463019723
